moved code to do with uvw coordinates and vis_params to separate header and impl files
