+++ cpp/purify/CMakeLists.txt
@@ -28,12 +28,13 @@ set(HEADERS
   wide_field_utilities.h
   wkernel_integration.h
   wproj_operators.h
+  uvw_utilities.h
   "${PROJECT_BINARY_DIR}/include/purify/config.h")
 
 set(SOURCES utilities.cc pfitsio.cc
   kernels.cc wproj_utilities.cc operators.cc uvfits.cc yaml-parser.cc
-  read_measurements.cc distribute.cc integration.cc wide_field_utilities.cc wkernel_integration.cc 
-  wproj_operators.cc)
+  read_measurements.cc distribute.cc integration.cc wide_field_utilities.cc wkernel_integration.cc
+  wproj_operators.cc uvw_utilities.cc)
 
 if(TARGET casacore::ms)
   list(APPEND SOURCES casacore.cc)
